+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - DictKey.bankName parameter
 - Institution resource to allow query of institutions recognized by the Brazilian Central Bank for Pix and TED transactions
 - TaxPayment resource
+- DarfPayment resource
 
 ### Fixed
 - special characters in brcodePreview query